requirements.txt
Flask==2.0.1
fpdf==1.7.2
Rtree==0.9.7
Fiona==1.8.20
geopandas==0.10.2
pandas==1.3.3
plotly==5.1.0
Werkzeug==2.0.1
gunicorn==20.1.0
kaleido==0.2.1
guppy3==3.1.2
statsmodels==0.13.2